requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.10
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6 
pydantic==2.4.2
pydantic-settings==2.0.3
httpx==0.25.1
python-dotenv==1.0.0
email-validator==2.1.0.post1
aiofiles==23.2.1 
PyMuPDF==1.22.3 
redis==5.2.1
bcrypt==4.0.1
requests==2.31.0
cryptography==42.0.2